+++ R/AAA_imports.R
@@ -1,3 +1,2 @@
-#' @importFrom utils head tail read.table type.convert
-#' @import graphics
+
 donothing <- function() {}