+++ R/AAA_pomdp-package.R
@@ -10,7 +10,15 @@
 #' @docType package
 #' @name pomdp-package
 #' 
+#' @importFrom utils head tail read.table type.convert
+#' @import graphics
 #' @import Rcpp
 #' @import foreach
 #' @useDynLib pomdp, .registration=TRUE
 NULL
+
+.onAttach <-function(libname, pkgname) {
+  ### silence warning for no backend
+  if (!foreach::getDoParRegistered())
+    foreach::registerDoSEQ()
+}